moved docker functions in a separate file
